Adding better handling of query failures
On the first few runs I noticed timeouts when quering quad-8 after making 20-30k
queries in a short period. This puts in a half second sleep. This does mean the
total duration could be longer, however I had considered using a maximum number
of failures however it is more appealing to have a complete dataset than to skip
records because of local resolver issues.