Adding new notebook for putting analysis together
This is deliberately using the TF template from the `rticles` module as there's
been too much grief trying to get it to work inside the Docker container.
Instead, the intention is to use this notebook and port it later into the
correct format after the data has been analysed.

You will also note this is working off a "snapshot" that is not checked in, the
full complete dataset will replace it in due course but any reference to values
will be templated to avoid referring to incorrect values.